indic_speech_system/requirements.txt
# requirements.txt
torch
torchvision
torchaudio
transformers
datasets
accelerate
librosa
soundfile
scipy
gradio

# Bot dependencies
python-telegram-bot==20.7
flask
flask-cors
flask-socketio
python-dotenv
aiohttp
aiofiles
requests
websockets

# AI dependencies
openai-whisper
ollama
pandas

# Database
psycopg2-binary

# Audio processing
pydub
ffmpeg-python
